Render BookViewer component
Books is the array book is the current item and i is the index of the current item.

If the book.bookImages exists than BookViewer is rendered.
BookViewer receives; the array with images, an unique key based on i, and the title.

Added skeleton class to book images from component library. It's a loading image.