Component for table
Made a new component for the book tabel in book-list

I changed the <table>  to a <ul> with <li> items.

If a item is not available i used an OR operator.

Placed a <span> element in the <p> so the 2 values can be styled separately.

Styled the elements like in the design.

And added a @media for a responsive design.